chore(plan) make plan form more modular, to ease re-usage
