+++ src/hooks/plans/usePlanForm.tsx
@@ -1,9 +1,10 @@
 import { gql } from '@apollo/client'
-import _omit from 'lodash/omit'
+import { FormikProps, useFormik } from 'formik'
 import { useEffect, useMemo } from 'react'
 import { generatePath, useNavigate, useParams } from 'react-router-dom'
+import { number, object, string } from 'yup'
 
-import { PlanFormInput } from '~/components/plans/types'
+import { LocalChargeInput, PlanFormInput } from '~/components/plans/types'
 import {
   addToast,
   hasDefinedGQLError,
@@ -15,16 +16,23 @@ import {
 import { FORM_ERRORS_ENUM } from '~/core/constants/form'
 import { CUSTOMER_DETAILS_ROUTE, ERROR_404_ROUTE, PLANS_ROUTE } from '~/core/router'
 import { serializePlanInput } from '~/core/serializers'
+import { deserializeAmount } from '~/core/serializers/serializeAmount'
+import { chargeSchema } from '~/formValidation/chargeSchema'
 import {
+  CurrencyEnum,
   DeletePlanDialogFragmentDoc,
   EditPlanFragment,
   EditPlanFragmentDoc,
   LagoApiError,
+  PlanInterval,
   PlanItemFragmentDoc,
   useCreatePlanMutation,
   useGetSinglePlanQuery,
   useUpdatePlanMutation,
 } from '~/generated/graphql'
+import { getPropertyShape } from '~/pages/CreatePlan'
+
+import { useInternationalization } from '../core/useInternationalization'
 
 gql`
   query getSinglePlan($id: ID!) {
@@ -54,17 +62,19 @@ gql`
 export type PLAN_FORM_TYPE = keyof typeof PLAN_FORM_TYPE_ENUM
 
 export interface UsePlanFormReturn {
-  loading: boolean
-  type: keyof typeof PLAN_FORM_TYPE_ENUM
-  parentPlanName?: string
   errorCode?: string
-  plan?: Omit<EditPlanFragment, 'name' | 'code'> & { name?: string; code?: string }
-  onSave: (values: PlanFormInput) => Promise<void>
+  formikProps: FormikProps<PlanFormInput>
+  isEdition: boolean
+  loading: boolean
+  parentPlanName?: string
+  plan?: (Omit<EditPlanFragment, 'name' | 'code'> & { name?: string; code?: string }) | null
+  type: PLAN_FORM_TYPE
   onClose: () => void
 }
 
 export const usePlanForm: () => UsePlanFormReturn = () => {
   const navigate = useNavigate()
+  const { translate } = useInternationalization()
   const { id } = useParams()
   const { parentId, subscriptionInput, customerId, type: actionType } = useOverwritePlanVar()
   const { data, loading, error } = useGetSinglePlanQuery({
@@ -75,6 +85,89 @@ export const usePlanForm: () => UsePlanFormReturn = () => {
   const isOverride = actionType === 'override' && !!parentId
   const isDuplicate = actionType === 'duplicate' && !!parentId
   const type = !!id ? 'edition' : isDuplicate ? 'duplicate' : isOverride ? 'override' : 'creation'
+  const isEdition = type === PLAN_FORM_TYPE_ENUM.edition
+  const plan = data?.plan
+  const shouldOffuscateForDuplicateAndOverride =
+    type === PLAN_FORM_TYPE_ENUM.override || type === PLAN_FORM_TYPE_ENUM.duplicate
+  const onSave =
+    type === PLAN_FORM_TYPE_ENUM.edition
+      ? async (values: PlanFormInput) => {
+          await update({
+            variables: {
+              input: { id: id as string, ...serializePlanInput(values) },
+            },
+          })
+        }
+      : async (values: PlanFormInput) => {
+          await create({
+            variables: {
+              input: {
+                ...(type === PLAN_FORM_TYPE_ENUM.override ? { parentId } : {}),
+                ...serializePlanInput(values),
+              },
+            },
+          })
+        }
+
+  const formikProps = useFormik<PlanFormInput>({
+    initialValues: {
+      name: shouldOffuscateForDuplicateAndOverride ? '' : plan?.name || '',
+      code: shouldOffuscateForDuplicateAndOverride ? '' : plan?.code || '',
+      description: plan?.description || '',
+      interval: plan?.interval || PlanInterval.Monthly,
+      taxes: plan?.taxes || [],
+      payInAdvance: plan?.payInAdvance || false,
+      amountCents: isNaN(plan?.amountCents)
+        ? ''
+        : String(
+            deserializeAmount(plan?.amountCents || 0, plan?.amountCurrency || CurrencyEnum.Usd)
+          ),
+      amountCurrency: plan?.amountCurrency || CurrencyEnum.Usd,
+      trialPeriod:
+        plan?.trialPeriod === null || plan?.trialPeriod === undefined
+          ? isEdition
+            ? 0
+            : undefined
+          : plan?.trialPeriod,
+      billChargesMonthly: plan?.billChargesMonthly || undefined,
+      charges: plan?.charges
+        ? plan?.charges.map(
+            ({ taxes, properties, groupProperties, minAmountCents, payInAdvance, ...charge }) => ({
+              taxes: taxes || [],
+              minAmountCents: isNaN(minAmountCents)
+                ? undefined
+                : String(
+                    deserializeAmount(minAmountCents || 0, plan.amountCurrency || CurrencyEnum.Usd)
+                  ),
+              payInAdvance: payInAdvance || false,
+              properties: properties ? getPropertyShape(properties) : undefined,
+              groupProperties: groupProperties?.length
+                ? groupProperties?.map((prop) => {
+                    return {
+                      groupId: prop.groupId,
+                      values: getPropertyShape(prop.values),
+                    }
+                  })
+                : [],
+              ...charge,
+            })
+          )
+        : ([] as LocalChargeInput[]),
+    },
+    validationSchema: object().shape({
+      name: string().required(''),
+      code: string().required(''),
+      interval: string().required(''),
+      amountCents: string().required(''),
+      trialPeriod: number().typeError(translate('text_624ea7c29103fd010732ab7d')).nullable(),
+      amountCurrency: string().required(''),
+      charges: chargeSchema,
+    }),
+    enableReinitialize: true,
+    validateOnMount: true,
+    onSubmit: onSave,
+  })
+
   const [create, { error: createError }] = useCreatePlanMutation({
     context: { silentErrorCodes: [LagoApiError.UnprocessableEntity] },
     onCompleted({ createPlan }) {
@@ -118,12 +211,13 @@ export const usePlanForm: () => UsePlanFormReturn = () => {
     },
   })
 
-  useEffect(() => {
-    if (hasDefinedGQLError('NotFound', error, 'plan')) {
-      navigate(ERROR_404_ROUTE)
+  const errorCode = useMemo(() => {
+    if (hasDefinedGQLError('ValueAlreadyExist', createError || updateError)) {
+      return FORM_ERRORS_ENUM.existingCode
     }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [error])
+
+    return undefined
+  }, [createError, updateError])
 
   // Clear duplicate plan var when leaving the page
   useEffect(() => {
@@ -135,43 +229,49 @@ export const usePlanForm: () => UsePlanFormReturn = () => {
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [])
 
-  const errorCode = useMemo(() => {
-    if (hasDefinedGQLError('ValueAlreadyExist', createError || updateError)) {
-      return FORM_ERRORS_ENUM.existingCode
+  useEffect(() => {
+    if (hasDefinedGQLError('NotFound', error, 'plan')) {
+      navigate(ERROR_404_ROUTE)
     }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [error])
 
-    return undefined
-  }, [createError, updateError])
+  useEffect(() => {
+    if (errorCode === FORM_ERRORS_ENUM.existingCode) {
+      formikProps.setFieldError('code', 'text_632a2d437e341dcc76817556')
+      const rootElement = document.getElementById('root')
+
+      if (!rootElement) return
+      rootElement.scrollTo({ top: 0 })
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [errorCode])
+
+  useEffect(() => {
+    if (
+      (!formikProps.values.charges ||
+        !formikProps.values.charges.length ||
+        formikProps.values.interval !== PlanInterval.Yearly) &&
+      !!formikProps.values.billChargesMonthly
+    ) {
+      formikProps.setFieldValue('billChargesMonthly', false)
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [
+    formikProps.values.charges,
+    formikProps.values.billChargesMonthly,
+    formikProps.values.interval,
+  ])
 
   return useMemo(
     () => ({
+      errorCode,
+      formikProps,
+      isEdition,
       loading,
       type,
-      errorCode,
+      plan,
       parentPlanName: data?.plan?.name,
-      plan:
-        type === PLAN_FORM_TYPE_ENUM.override || type === PLAN_FORM_TYPE_ENUM.duplicate
-          ? _omit(data?.plan || undefined, ['name', 'code'])
-          : data?.plan || undefined,
-      onSave:
-        type === PLAN_FORM_TYPE_ENUM.edition
-          ? async (values) => {
-              await update({
-                variables: {
-                  input: { id: id as string, ...serializePlanInput(values) },
-                },
-              })
-            }
-          : async (values) => {
-              await create({
-                variables: {
-                  input: {
-                    ...(type === PLAN_FORM_TYPE_ENUM.override ? { parentId } : {}),
-                    ...serializePlanInput(values),
-                  },
-                },
-              })
-            },
       onClose: () => {
         if (type === PLAN_FORM_TYPE_ENUM.override) {
           navigate(generatePath(CUSTOMER_DETAILS_ROUTE, { id: customerId as string }))
@@ -180,7 +280,6 @@ export const usePlanForm: () => UsePlanFormReturn = () => {
         }
       },
     }),
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-    [loading, id, customerId, parentId, type, data?.plan, errorCode, update, create]
+    [errorCode, formikProps, isEdition, loading, type, plan, data?.plan?.name, navigate, customerId]
   )
 }
